refactor: migrate hooks to services
