+++ app/admin/compliance.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text, StyleSheet } from 'react-native';
 import { useTheme } from '@/ui/ThemeProvider';
-import useRequirePlatformAdmin from '../../hooks/useRequirePlatformAdmin';
+import { useRequirePlatformAdmin } from '@/services';
 import RequireWallet from '../../components/RequireWallet';
 
 export default function Compliance() {
