+++ app/admin/settings.tsx
@@ -2,9 +2,9 @@ import { errorLog } from '@/utils/logger';
 import React, { useState, useEffect } from 'react';
 import { View, Text, StyleSheet, TouchableOpacity, ScrollView } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import useAppRouter from 'hooks/useAppRouter';
+import { useAppRouter } from '@/services';
 import { ArrowLeft, Save, Settings as SettingsIcon } from 'lucide-react-native';
-import useRequirePlatformAdmin from '../../hooks/useRequirePlatformAdmin';
+import { useRequirePlatformAdmin } from '@/services';
 import { useTheme } from '@/ui/ThemeProvider';
 import { useCurrency } from '../../contexts/CurrencyContext';
 import { useLanguage } from '@/ui/ThemeProvider';
