+++ app/admin/user-directory.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { View, Text, StyleSheet } from 'react-native';
 import { useTheme } from '@/ui/ThemeProvider';
-import useRequirePlatformAdmin from 'hooks/useRequirePlatformAdmin';
+import { useRequirePlatformAdmin } from '@/services';
 import RequireWallet from '@/components/RequireWallet';
 import chain from '@/services/chain';
 import { User } from '@/types';
