+++ app/categories.tsx
@@ -3,7 +3,7 @@ import { ScrollView, TouchableOpacity, Text, StyleSheet } from 'react-native';
 import AppShell from '../components/layout/AppShell';
 import { useTheme } from '@/ui/ThemeProvider';
 import { useLanguage } from '@/ui/ThemeProvider';
-import useAppRouter from 'hooks/useAppRouter';
+import { useAppRouter } from '@/services';
 import EmptyState from '@/shared/ui/EmptyState';
 import { Plus } from 'lucide-react-native';
 import ErrorBoundary from '@/shared/ErrorBoundary';
