+++ app/store/[storeId]/admin/_BulkUploadScreen.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { Alert } from 'react-native';
-import useAppRouter from 'hooks/useAppRouter';
+import { useAppRouter } from '@/services';
 import AdminBulkUploader from '@/features/products/components/AdminBulkUploader';
 import { useAuth } from '@/features/auth/AuthContext';
