+++ app/store/[storeId]/admin/_PricingTiersScreen.tsx
@@ -9,7 +9,7 @@ import {
   Platform,
 } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import useAppRouter from 'hooks/useAppRouter';
+import { useAppRouter } from '@/services';
 import {
   ArrowLeft,
   Plus,
