--- hooks/useFrameworkReady.ts
@@ -1,13 +1,0 @@
-import { useEffect } from 'react';
-
-declare global {
-  interface Window {
-    frameworkReady?: () => void;
-  }
-}
-
-export function useFrameworkReady() {
-  useEffect(() => {
-    window.frameworkReady?.();
-  });
-}
