+++ tsconfig.json
@@ -69,9 +69,6 @@
       "@/*": [
         "./*"
       ],
-      "hooks/*": [
-        "hooks/*"
-      ],
       "@expo/metro-runtime/src/HMRClient": [
         "./HMRClient.ts"
       ],
